+++ examples/docker-compose/functions/sub/index.ts
@@ -1,0 +1,3 @@
+export default (req, res) => {
+  res.status(200).send(`Index function in a sub-directory`)
+}