+++ examples/react-apollo/src/components/AuthLink.tsx
@@ -1,38 +1,41 @@
 import { Link } from 'react-router-dom'
 
-import { Button, ButtonVariant } from '@mantine/core'
+import { Button, ButtonProps, SharedButtonProps } from '@mantine/core'
 
-const AuthLink: React.FC<{
-  icon?: React.ReactNode
-  link: string
-  color?: string
-  children?: React.ReactNode
-  variant?: ButtonVariant
-}> = ({ icon, color, link, variant, children }) => {
-  return (
-    <Button
-      role="button"
-      component={Link}
-      fullWidth
-      radius="sm"
-      variant={variant}
-      to={link}
-      leftIcon={icon}
-      styles={(theme) => ({
-        root: {
-          backgroundColor: color,
-          '&:hover': {
-            backgroundColor: color && theme.fn.darken(color, 0.05)
-          }
-        },
-
-        leftIcon: {
-          marginRight: 15
+const AuthButton: <C = 'button'>(props: ButtonProps<C>) => React.ReactElement = ({
+  color,
+  ...rest
+}) => (
+  <Button
+    role="button"
+    fullWidth
+    radius="sm"
+    styles={(theme) => ({
+      root: {
+        backgroundColor: color,
+        '&:hover': {
+          backgroundColor: color && theme.fn.darken(color, 0.05)
         }
-      })}
-    >
-      {children}
-    </Button>
+      },
+
+      leftIcon: {
+        marginRight: 15
+      }
+    })}
+    {...rest}
+  />
+)
+
+const AuthLink: React.FC<
+  SharedButtonProps & {
+    link: string
+  }
+> = ({ link, ...rest }) => {
+  const isExternal = link.startsWith('http://') || link.startsWith('https://')
+  return isExternal ? (
+    <AuthButton component={'a'} href={link} {...rest} />
+  ) : (
+    <AuthButton component={Link} to={link} {...rest} />
   )
 }