+++ examples/vue-apollo/src/pages/sign-in/CommonActions.vue
@@ -11,13 +11,7 @@
   >
     Continue with passwordless email
   </v-btn>
-  <v-btn
-    class="my-1"
-    block
-    variant="text"
-    color="primary"
-    to="/signin/email-password"
-  >
+  <v-btn class="my-1" block variant="text" color="primary" to="/signin/email-password">
     Continue with email + password
   </v-btn>
 </template>