+++ packages/hasura-storage-js/tests/helpers/config.ts
@@ -1,0 +1,4 @@
+/**
+ * Base Nhost backend url for the Nhost client.
+ */
+export const BASE_URL = 'http://localhost:1337'