+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @nhost/nextjs
 
+## 1.6.1
+
+### Patch Changes
+
+- Updated dependencies [6eaa5c79]
+  - @nhost/core@0.7.4
+  - @nhost/react@0.11.1
+  - @nhost/nhost-js@1.4.5
+
+## 1.6.0
+
+### Patch Changes
+
+- Updated dependencies [4f928756]
+  - @nhost/react@0.11.0
+  - @nhost/nhost-js@1.4.4
+
+## 1.5.0
+
+### Patch Changes
+
+- Updated dependencies [f9854b15]
+- Updated dependencies [f9854b15]
+  - @nhost/react@0.10.0
+  - @nhost/core@0.7.3
+  - @nhost/nhost-js@1.4.3
+
+## 1.4.2
+
+### Patch Changes
+
+- dbc10e62: fixed `exports` field to support imports in a server-side environment
+- Updated dependencies [dbc10e62]
+  - @nhost/core@0.7.2
+  - @nhost/nhost-js@1.4.2
+  - @nhost/react@0.9.2
+
 ## 1.4.1
 
 ### Patch Changes